refactor time millis to time seconds and move to time utils
