+++ app/src/main/java/app/stacq/plan/util/time/TimeUtil.kt
@@ -1,7 +1,11 @@
 package app.stacq.plan.util.time
 
 import android.os.SystemClock
+import app.stacq.plan.util.constants.TimerConstants
 import java.time.Instant
+import java.time.LocalDateTime
+import java.time.LocalTime
+import java.time.ZoneOffset
 
 class TimeUtil {
 
@@ -12,11 +16,6 @@ class TimeUtil {
         return (finishAt - now) * 1000L
     }
 
-    fun secondsInFuture(finishAt: Long): Long {
-        val now: Long = instant.epochSecond
-        return ((finishAt / 1000L) - now)
-    }
-
     fun alarmTriggerTimer(finishAt: Long): Long {
         val now: Long = instant.epochSecond
         val futureMillis = (finishAt - now) * 1000L
@@ -24,6 +23,29 @@ class TimeUtil {
     }
 
     fun plusSecondsEpoch(seconds: Long): Long {
-        return instant.plusSeconds(seconds).epochSecond
+        return instant.plusMillis(seconds).epochSecond
+    }
+
+    fun timeSeconds(hour: Int, minute: Int): Long {
+        val now = LocalDateTime.now()
+        val timerTime = LocalTime.of(hour, minute)
+        val timerDateTime = LocalDateTime.of(now.toLocalDate(), timerTime)
+
+        val millis = if (timerDateTime.isAfter(now)) {
+            // timer time on clock is in future
+            val timerEpochTime =
+                timerDateTime.atZone(ZoneOffset.UTC).toInstant().toEpochMilli()
+            val nowEpochTime = now.atZone(ZoneOffset.UTC).toInstant().toEpochMilli()
+            timerEpochTime - nowEpochTime
+        } else {
+            // timer time on clock is past
+            val tomorrowDateTime = timerDateTime.plusDays(1)
+            val timerEpochTime =
+                tomorrowDateTime.atZone(ZoneOffset.UTC).toInstant().toEpochMilli()
+            val nowEpochTime = now.atZone(ZoneOffset.UTC).toInstant().toEpochMilli()
+            timerEpochTime - nowEpochTime
+        }
+
+        return millis / TimerConstants.TIME_MILLIS_TO_SECONDS
     }
 }
